requirements.txt
numpy==1.24.3
pandas==1.4.1
geopandas==0.13.2
tdqm
joblib==1.2.0
scikit-learn==1.2.2
xgboost==1.6.2
lightgbm==3.3.5
haversine
pyspark==3.3.2
apache-airflow[postgres]==2.6.2
python-dateutil
pyarrow==13.0.0